Updated model
Updated model to fetch totals information for covid data from worldometers. Updates the data class which the lambda uses to send the data to the frontend. The class now returns total data on start for all countries as well as the individual countries geolocation to be mapped. When a country is clicked on the map, a new fetch request can be sent with the 'switch': <country name> header to receive covid data related to that country.