Flask API converted to Lambda API + Performance Changes
Image Compression + Iconify Removed for performance reasons. Flask API converted to Lambda API as it is able to be deployed for free this way. The Flask API is still in the project folder as it is up to date and can be used for localhost testing/development and it will be put on a server at some point to replace the lambda api. The lambda API was implemented mainly because it was quick.